lfs,commands: remove lfs.NewUploadable() in favor of commands.uploadTransfer
