+++ lfs/upload_queue.go
@@ -1,78 +1,9 @@
 package lfs
 
 import (
-	"fmt"
-	"os"
-	"path/filepath"
-
 	"github.com/git-lfs/git-lfs/config"
-	"github.com/git-lfs/git-lfs/errors"
 	"github.com/git-lfs/git-lfs/tq"
 )
-
-// NewUploadable builds the Uploadable from the given information.
-// "filename" can be empty if a raw object is pushed (see "object-id" flag in push command)/
-func NewUploadable(oid, filename string) (*tq.Transfer, error) {
-	localMediaPath, err := LocalMediaPath(oid)
-	if err != nil {
-		return nil, errors.Wrapf(err, "Error uploading file %s (%s)", filename, oid)
-	}
-
-	if len(filename) > 0 {
-		if err := ensureFile(filename, localMediaPath); err != nil {
-			return nil, err
-		}
-	}
-
-	fi, err := os.Stat(localMediaPath)
-	if err != nil {
-		return nil, errors.Wrapf(err, "Error uploading file %s (%s)", filename, oid)
-	}
-
-	return &tq.Transfer{
-		Name: filename,
-		Oid:  oid,
-		Size: fi.Size(),
-		Path: localMediaPath,
-	}, nil
-}
-
-// ensureFile makes sure that the cleanPath exists before pushing it.  If it
-// does not exist, it attempts to clean it by reading the file at smudgePath.
-func ensureFile(smudgePath, cleanPath string) error {
-	if _, err := os.Stat(cleanPath); err == nil {
-		return nil
-	}
-
-	expectedOid := filepath.Base(cleanPath)
-	localPath := filepath.Join(config.LocalWorkingDir, smudgePath)
-	file, err := os.Open(localPath)
-	if err != nil {
-		return err
-	}
-
-	defer file.Close()
-
-	stat, err := file.Stat()
-	if err != nil {
-		return err
-	}
-
-	cleaned, err := PointerClean(file, file.Name(), stat.Size(), nil)
-	if cleaned != nil {
-		cleaned.Teardown()
-	}
-
-	if err != nil {
-		return err
-	}
-
-	if expectedOid != cleaned.Oid {
-		return fmt.Errorf("Trying to push %q with OID %s.\nNot found in %s.", smudgePath, expectedOid, filepath.Dir(cleanPath))
-	}
-
-	return nil
-}
 
 // NewUploadQueue builds an UploadQueue, allowing `workers` concurrent uploads.
 func NewUploadQueue(cfg *config.Configuration, options ...tq.Option) *tq.TransferQueue {
